chore: move SummaryGrid to @molecules
