+++ src/components/molecules/index.ts
@@ -37,3 +37,4 @@ export {RunningContextType} from './RunningContext';
 export {default as MessagePanel} from './MessagePanel';
 export {default as InlineNotification} from './InlineNotification';
 export {default as AiInsightsTab} from './AiInsightsTab';
+export {default as SummaryGrid} from './SummaryGrid';
